requirements.txt
streamlit
requests
pdfminer.six
PyPDF2
